Much better, works now.
Has
- AppErrorBoundary.tsx
- And "app alert" in AppContext
