+++ api/Metrix.Api/Source/Controllers/MetricsController.cs
@@ -22,8 +22,6 @@ public class MetricsController : ControllerBase
   [HttpGet]
   public Metric[] GetAll()
   {
-    throw new Exception("Fuuck off!");
-    
     return _dispatcher.Query(new GetAllMetricsQuery());
   }
